requirements.txt
# ── Core Server ──────────────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic

# ── LLM: Google Gemini ────────────────────────────────────────
google-genai

# ── LangChain / LangGraph ─────────────────────────────────────
langchain
langchain-chroma
langchain-text-splitters
langchain-huggingface
langgraph

# ── RAG / Vector Store ────────────────────────────────────────
chromadb
sentence-transformers
huggingface_hub

# ── Document Loading ──────────────────────────────────────────
pdfplumber
rich

# ── Voice (optional — uncomment if needed) ───────────────────
# pyttsx3
# elevenlabs
# openai-whisper
